Rapier physics (#27)
This commit introduces the rapier-physics engine with the focus on the architecture. The integration does not yet provide all functionalities Rapier has not offer, but this will be fixed in future commits.

The physics-engine runs in it's own thread and the user can register colliders, character controllers and rigid-bodies via `PhysicsWorkerContext`. Method calls are transformed into `PhysicsWorkerTask`, sent to the physics worker thread and executed there in the order they arrived.

All properties that are shared between the proxy and it's origin are stored in a triple buffer and updated after each `step` call.